feat: change trainer pilot session logic
Previously, when a training Rig tried to pilot for the first time, its session was simply updated. This was misleading because changing from a trainer to a pilot means the session ended, and it also impacted how pilot history was captured. Namely, if the row is updated and maintains the training session, and once a real pilot is specified, it looks as though the pilot was piloting the Rig the entire time. Training history was lost. The new logic simply ends the trainer session and starts a new pilot session at the same time, if the Rig is currently in-flight while trained and trying to pilot for the first time.