fix: update verify script error handling
Verifying contracts uses hardhat + etherscan. If a contract is already verified and you attempt to verify it again, a halting error is thrown such that any new contracts cannot be verified. This fix adds logic to handle the verification-specific errors and ignore them, allowing already verified and unverified contracts to coexist in the same script.