feat: token reputation This PR implements token-based reputation. It is inclusive of a draft (unsubmitted) EIP located in `research`, the minimal `ITokenReputation` interface, and associated adjustments to other contracts and tests.
In relation to the EIP, the following provides more detail and can be found at `research/eip-draft_token_reputation.md`:
- This is the core EIP for token-based reputation. It is a minimal interface that simply has staking & unstaking methods along with events; the design is to be as lightweight as possible and hone in on metadata having the "value" and storing reputation using indexed events / data availability.
- With a goal of EIP acceptance, the majority of the EIP focuses on non-Tableland specific features and includes real-world usage by OpenSea today. [OS disables transfers](https://docs.opensea.io/docs/metadata-standards#disable-trading-for-staked-or-locked-tokens) on a `Stake` event as long as it includes a `tokenId` parameter; it can include additional params and still works as expected (according to the OS support team). The `Unstake` event will re-enable sales.
- At the end of the EIP, two example implementations are provided. The first is generic but at least names the ERC721 token and `baseURI` with Tableland.
- The second example is very Tableland specific (SQLHelpers, TablelandDeployments, gateway query) and tries to sort of generalize the example using the comments. This example is fully functional (you can deploy the code) but, perhaps, isn't the "most real" implementation; however, an EIP must be as generic as possible, so this was the ultimate idea with the attempt.

With the `ITokenReputation.sol` contract, it has been implemented by the `TablelandRigs.sol` contract. This was easier than doing so in `TablelandPilots.sol` as the function signature was more aligned to the core Rigs contract. It was rather straightforward to drop it in, which is a good sign from a usability standpoint (only took a few mins to adjust the exist train / pilot / park methods and tests). Note that this does not change any core contract logic; only external apps like The Garage are affected since the Rigs Garage functions are now different.